Use version compare instead of `curl || curl`
The idea had promise but the first curl ends up not returning a non-zero
exit code even when the status code is a 404. Which means we'd have to
inspect the contents of the file. To do that we have to rely on
consistency of tools across platforms (does `file` return the same
output on all platforms? Could we use `gunzip -t` to test? Does that
exist on all systes?).

Instead, we opt for the simple (but admitedly brittle) version check and
cross our fingers that eksctl doesn't change again.